+++ django/cart/tests.py
@@ -2,7 +2,7 @@ from django.test import TestCase
 # from django.core.management import call_command
 # from django.core.management.base import CommandError
 # from django.core.exceptions import ValidationError
-from django.core.urlresolvers import reverse
+from django.urls import reverse
 
 # cart.py
 # write test for adding to cart
