+++ django/inventory/models.py
@@ -1,4 +1,4 @@
-from django.core.urlresolvers import reverse
+from django.urls import reverse
 from django.db import models
 from meadery.models import Product
 from django.core.exceptions import ValidationError
