+++ django/utils/tests.py
@@ -2,7 +2,7 @@
 # from django.core.management import call_command
 # from django.core.management.base import CommandError
 # from django.core.exceptions import ValidationError
-# from django.core.urlresolvers import reverse
+# from django.urls import reverse
 
 # modeladmin.py
 # SmarterModelAdmin
